Use single quotes if the path contains double quotes
If there is a quote in the folder name, the mkdir step fails as it's
contained in double quotes.
What I was thinking is that for all special characters apart from double
quotes, we can enclose the folder path with double quotes.
For double quotes, we enclose the path in single quotes.