Add entry point for running the Lambda script locally
Add a Scala App object which calls the Lambda, so that you can
run the code locally. If you set the environment variables to
point to the queues on the integration environment, it will
download the file from intg S3 and send messages to the intg SQS
queues.

It saves the S3 file to a local folder instead of EFS, because
EFS volumes are mounted on Lambda as if they were local folders
anyway.

The code for generating an SNS message is quite verbose because
there's no easy way to serialize an SNS object to JSON. The AWS
SDK v1 has a class called `SnsMessageManager` which might help,
but it's not available in the SDK v2.