Fix NullPointerException when message body has wrong format
If the message body could not be parsed as an S3 event, Circe would
throw an exception. This exception has no nested cause, so the call to
`getCause` in the error handling was causing a NullPointerException.

This meant that the Lambda would fail before the successful messages
were deleted from the queue, causing them all to be retried. It also
hides the original cause of the error.

This is not a big problem at the moment because the batch size is 1, but
it would be a problem if we ever did start processing multiple messages.