Reset message visibility if an error happens when downloading the files
This makes Lambda retries faster. They had become much slower when we
increased the SQS message visibility to be three times the length of the
Lambda timeout: https://github.com/nationalarchives/tdr-dev-documentation/blob/master/architecture-decision-records/0020-sqs-visibility-timeout.md

This change means that the retry happens immediately, not when the message
has expired after a few minutes.

The message visibility reset only happens if an error is encountered when
fetching the original filepath from the API, downloading the file or updating
the downstream queues. Those types of errors might be caused by temporary
network glitches, and can be retried.