Send userId and dirty bucket name
This is to allow the antivirus checks to work with both the existing
dirty bucket and the new cloudfront dirty bucket so we pass the bucket
that triggered the notification to the antivirus lambda.

I've also removed cognitoId from the messages sent to the checksum
lambda and file format lambda as neither lambda uses them.

I've added userId alongside cognitoId so the antivirus doesn't break
while deploying this change.

This PR needs to deploy first, then
https://github.com/nationalarchives/tdr-antivirus/pull/37 and when
that's merged, cognitoId can be removed from the antivirus output in
this lambda.